services/runtime/apps/api/pyproject.toml
[project]
name = "pfios-api"
version = "0.1.0"
description = "Financial-AI-OS API Service"
authors = [{ name = "User" }]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "duckdb>=0.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "ccxt>=4.0.0",
    "pandas>=2.0.0",
    "httpx>=0.24.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
